add login btn as a guest and rm file dont need
